move database project shared files to shared database project
