+++ Database/GlobalMapper.cs
@@ -9,14 +9,5 @@ public class GlobalMapper : Profile
 
         RecognizeDestinationPostfixes("Id");
         RecognizePostfixes("Id");
-
-        CreateMap<Money, decimal>()
-            .ConvertUsing(src => src.Value);
-        CreateMap<Money, decimal?>()
-            .ConvertUsing(src => src != null ? src.Value : null);
-        CreateMap<EntityReference, Guid>()
-            .ConvertUsing(src => src.Id);
-        CreateMap<EntityReference, Guid?>()
-            .ConvertUsing(src => src.Id);
     }
 }
